taqueriataco/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
djangorestframework>=3.14,<4.0
